moves docker build out of js api
